feat(perps): rename PerpsSection to Perpetuals and implement new component structure
