+++ app/components/Views/Homepage/Sections/Perpetuals/components/PerpsPositionRow/PerpsPositionRow.styles.ts
